Significantly improved acquisition, waveform and image creation speed (#66)
* Significantly improved Acquisition, waveform and image creation speed, resulting in faster deserialization

* Updated version

+++ tests/test_waveform.py
@@ -17,8 +17,7 @@ def test_initialization_from_array():
     nchannels = 32
     nsamples = 256
 
-    data = numpy.random.randint(0, 1 << 32, size=(nchannels, nsamples))
-    data = data.astype(np.uint32)
+    data = numpy.random.randint(0, 1 << 32, size=(nchannels, nsamples),dtype=np.uint32)
 
     waveform = ismrmrd.Waveform.from_array(data)
